Refrain from watching from the start of time
syncthing-activity needs not be running 24/7 like Syncthing should be. For that,
it should only start tracking the instance when it starts up, not the start of
time. For that, the first thing it does is to poll for all the events and
retrieves the latest event ID. syncthing-activity will then go ahead to track
all events that happen AFTER that.

Signed-off-by: Arif Er <arifer612@pm.me>