GitHub Actions for linting and README building
Making use of source block executables in Org files, the README file will be
built on every release by means of CI/CD. The Python script will be linted by
flake8 through CI/CD as well.

The `__status__` metadata will be determined by the contents of the VERSION
file, i.e. on the master branch where the semantic versioning takes the form of
v*.*.* with no hyphens unlike the development branches that will have either
'-rcX' or '-betaX', it is understood to be in production.

Resolves #5

Signed-off-by: Arif Er <arifer612@pm.me>